Expose secure-random-algorithm configuration property
Expose secure-random-algorithm configuration property for both, http client and http server.

On Linux (Unix) based operation systems NativePRNG algorithm is being used. This algorithm
fetched random bytes directly from the operation system using the /dev/random and /dev/urandom
files. It uses /dev/random to get the seed, and /dev/urandom to get the bytes.

The problem is that /dev/urandom is still slow, and single threaded.

Below is the test on the machine with 4 CPU

  for i in {1..}; do dd if=/dev/urandom of=/dev/null bs=8K count=3000 & done

  3000+0 records in
  3000+0 records out
  24576000 bytes (25 MB, 23 MiB) copied, 1,81277 s, 8,9 MB/s

  for i in {1..4}; do dd if=/dev/urandom of=/dev/null bs=8K count=3000 & done

  3000+0 records in
  3000+0 records out
  24576000 bytes (25 MB, 23 MiB) copied, 6,39533 s, 2.2 MB/s
  3000+0 records in
  3000+0 records out
  24576000 bytes (25 MB, 23 MiB) copied, 6,49464 s, 2,2 MB/s
  3000+0 records in
  3000+0 records out
  24576000 bytes (25 MB, 23 MiB) copied, 6,55815 s, 2,3 MB/s
  3000+0 records in
  3000+0 records out
  24576000 bytes (25 MB, 23 MiB) copied, 6,60145 s, 2,1 MB/s

HTTPS (TLS) with block encryption requires ~16 bytes of random data for every single
encrypted block which is ~16kb of size. When the network is fast (infini-band) than
given throughput of /dev/urandom could be a bottle neck. For such cases it is better
to have a way how to switch the random number generator to something less strict, but
still secure. e.g.: java based SHA1PRNG generator that takes seed from the system,
but uses it's own long enough sequence to generate random bytes.